edit details modal is separate component, so it can be re-used
